+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Decoupling pool factory from reference pool deployment
+-   Allow pool factory to change reference pool implementation
+
 ## [1.0.0-alpha.2] - 2021-06-30
 
 ### Changed